requirements.txt
fbchat==1.4.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
py-trello==0.15.0
pydotenv==0.0.7
